@@ -3,53 +3,53 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-regular-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@material-ui/core": "^4.12.2",
-    "@material-ui/styles": "^4.11.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^16.4.3",
-    "@types/react": "^17.0.15",
-    "@types/react-dom": "^17.0.9",
-    "axios": "^0.21.1",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@fortawesome/fontawesome-svg-core": "^6.7.2",
+    "@fortawesome/free-brands-svg-icons": "^6.7.2",
+    "@fortawesome/free-regular-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@mui/material": "^6.3.1",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^22.10.5",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
+    "axios": "^1.7.9",
     "axios-cache-adapter": "^2.7.3",
-    "babel-eslint": "^10.1.0",
-    "color": "3.2.1",
-    "debug": "^4.3.2",
-    "localforage": "^1.9.0",
+    "color": "^4.2.3",
+    "debug": "^4.4.0",
+    "localforage": "^1.10.0",
     "localforage-memoryStorageDriver": "^0.9.2",
     "lodash": "^4.17.21",
-    "moment": "^2.29.1",
-    "node-sass": "4.14.1",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "react-transition-group": "^4.4.2",
-    "sweetalert2": "^11.1.2",
+    "moment": "^2.30.1",
+    "prop-types": "^15.8.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-helmet-async": "^2.0.5",
+    "react-router-dom": "^6.29.0",
+    "react-scripts": "^5.0.1",
+    "react-transition-group": "^4.4.5",
+    "sass": "^1.83.4",
+    "sweetalert2": "^11.15.2",
     "toastr": "^2.1.4",
-    "typescript": "^4.3.5",
-    "web-vitals": "^1.0.1"
+    "typescript": "^5.7.3",
+    "web-vitals": "^4.2.4"
   },
   "scripts": {
-    "start": "yarn build && react-scripts start",
-    "build": "tsc && react-scripts build",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
       "react-app",
-      "react-app/jest",
-      "plugin:react-hooks/recommended"
+      "react-app/jest"
     ]
   },
   "browserslist": {
@@ -65,17 +65,17 @@
     ]
   },
   "devDependencies": {
-    "@types/color": "^3.0.2",
-    "@types/debug": "^4.1.7",
-    "@types/react-helmet": "^6.1.2",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/toastr": "^2.1.39",
-    "@typescript-eslint/parser": "^4.28.5",
-    "eslint": "^7.31.0",
-    "eslint-config-react": "^1.1.7",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "prettier": "^2.3.2"
+    "@types/color": "^3.0.6",
+    "@types/debug": "^4.1.12",
+    "@types/lodash": "^4.17.23",
+    "@types/react-helmet-async": "^1.0.3",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/toastr": "^2.1.43",
+    "@typescript-eslint/parser": "^8.19.1",
+    "eslint": "^9.18.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "prettier": "^3.4.2"
   },
   "proxy": "http://localhost:5000"
 }
