@@ -1,48 +1,43 @@
 {
   "dependencies": {
-    "body-parser": "^1.19.0",
-    "cookie-parser": "^1.4.5",
+    "body-parser": "^1.20.3",
+    "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "node": "12.16.1",
-    "path": "^0.12.7",
-    "querystring": "^0.2.1",
-    "request": "^2.88.2",
+    "dotenv": "^16.4.7",
+    "express": "^4.21.2",
     "spotify-web-api-node": "^5.0.2"
   },
   "name": "spotify_playlist_generator_node",
   "version": "1.0.0",
   "main": "dist/server.js",
   "devDependencies": {
-    "@types/cookie-parser": "^1.4.2",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/fs-extra": "^9.0.12",
-    "@types/lodash": "^4.14.171",
-    "@types/node": "^16.4.3",
-    "@types/request": "^2.48.6",
-    "@types/shelljs": "^0.8.9",
+    "@eslint/js": "^9.18.0",
+    "@types/body-parser": "^1.19.5",
+    "@types/cookie-parser": "^1.4.7",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^5.0.0",
+    "@types/node": "^22.10.5",
     "@types/spotify-web-api-node": "^5.0.3",
-    "concurrently": "^6.2.0",
-    "dotenv": "^10.0.0",
-    "eslint": "^7.31.0",
-    "fs-extra": "^10.0.0",
-    "nodemon": "^2.0.12",
-    "npm-run-all": "^4.1.5",
-    "prettier": "^2.3.2",
-    "rimraf": "^3.0.2",
-    "shelljs": "^0.8.4",
-    "ts-node": "^10.1.0",
-    "tslint": "^6.1.3",
-    "typescript": "^4.3.5"
+    "@typescript-eslint/eslint-plugin": "^8.19.1",
+    "@typescript-eslint/parser": "^8.19.1",
+    "concurrently": "^9.1.2",
+    "eslint": "^9.18.0",
+    "nodemon": "^3.1.9",
+    "prettier": "^3.4.2",
+    "rimraf": "^6.0.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.7.3"
   },
   "scripts": {
+    "bootstrap": "./script/bootstrap",
+    "start:dev": "./script/start",
+    "start:prod": "./script/start --production",
     "clean": "rimraf dist/*",
     "copy-assets": "ts-node tools/copyAssets",
-    "prebuild": "tslint -c tslint.json -p tsconfig.json --fix",
-    "build": "yarn clean && yarn lint && yarn tsc && yarn copy-assets",
+    "prebuild": "yarn lint",
+    "build": "yarn clean && yarn lint && yarn tsc",
     "prestart": "yarn run build",
-    "lint": "tslint -c tslint.json -p tsconfig.json --fix",
+    "lint": "eslint . --ext .ts --fix",
     "tsc": "tsc",
     "client": "cd client && yarn start",
     "server": "node .",
