@@ -1,10 +1,10 @@
 'use-strict';
-import { Container, LinearProgress } from '@material-ui/core';
-import { ThemeProvider } from '@material-ui/styles';
+import { Container, LinearProgress } from '@mui/material';
+import { ThemeProvider } from '@mui/material/styles';
 import PropTypes from 'prop-types';
 import React, { lazy, Suspense, useCallback, useContext, useEffect } from 'react';
-import { Helmet } from 'react-helmet';
-import { Redirect, Route, Switch, withRouter } from 'react-router-dom';
+import { Helmet } from 'react-helmet-async';
+import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 import { TransitionGroup } from 'react-transition-group';
 import { logger } from './common-util';
 import { FadeIn } from './components/common-ui';
@@ -18,9 +18,9 @@ import { theme } from './styles/theme';
 const Drawer = lazy(() => import('./components/drawer/Drawer'));
 const OptionsModal = lazy(() => import('./components/common-ui/OptionsModal'));
 
-const App = (props: any) => {
+const App = () => {
     const { state, dispatch } = useContext(AppCtx);
-    const { location } = props;
+    const location = useLocation();
 
     const fetchMe = useCallback(async () => {
         try {
@@ -63,24 +63,26 @@ const App = (props: any) => {
 
         const redirect = !state.isAuthenticated ? '/welcome' : '/dashboard';
 
-        const routeComps = filteredRoutes.map((route) => {
+        const routeComps = filteredRoutes.flatMap((route) => {
             const Comp = route.component;
-            return (
-                <Route exact={route.exact} path={route.paths} key={route.id}>
-                    {(routerProps) => (
+            return route.paths.map((path, index) => (
+                <Route
+                    path={path}
+                    key={`${route.id}-${index}`}
+                    element={
                         <TransitionGroup id="app-transition-group" key="app-transition-group">
                             <FadeIn in key={route.id}>
                                 <React.Fragment>
-                                    <Comp {...routerProps} />
+                                    <Comp />
                                 </React.Fragment>
                             </FadeIn>
                         </TransitionGroup>
-                    )}
-                </Route>
-            );
+                    }
+                />
+            ));
         });
 
-        const mappedRoutes = routeComps.concat([<Redirect key="redirect" to={redirect} />]);
+        const mappedRoutes = routeComps.concat([<Route key="redirect" path="*" element={<Navigate to={redirect} replace />} />]);
 
         return mappedRoutes;
     }, [state.isAuthenticated]);
@@ -106,11 +108,11 @@ const App = (props: any) => {
                     <title>festival.me</title>
                 </Helmet>
                 <Suspense fallback={<LinearProgress variant="buffer" color="primary" value={20} valueBuffer={50} />}>
-                    <NavBar {...props}>
+                    <NavBar>
                         <OptionsModal />
                         <Drawer />
                         <Container>
-                            <Switch>{mapRenderRoutes()}</Switch>
+                            <Routes>{mapRenderRoutes()}</Routes>
                         </Container>
                         {renderFooter()}
                     </NavBar>
@@ -124,4 +126,4 @@ App.propTypes = {
     props: PropTypes.shape({}),
 };
 
-export default withRouter(App);
+export default App;
