@@ -1,4 +1,4 @@
-import { Typography, useTheme } from '@material-ui/core';
+import { Typography, useTheme } from '@mui/material';
 import PropTypes from 'prop-types';
 import { useCallback, useContext, useEffect, useState } from 'react';
 import { capitalize, logger, usePrevious } from '../../common-util';
