@@ -1,20 +1,25 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import { HelmetProvider } from 'react-helmet-async';
 import App from './App';
 import { AppContextProvider } from './AppContextProvider';
 import './index.css';
 import { BrowserRouter as Router } from 'react-router-dom';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(
+    document.getElementById('root') as HTMLElement
+);
+root.render(
     <React.StrictMode>
-        <AppContextProvider>
-            <Router basename="/">
-                <App />
-            </Router>
-        </AppContextProvider>
-    </React.StrictMode>,
-    document.getElementById('root'),
+        <HelmetProvider>
+            <AppContextProvider>
+                <Router basename="/">
+                    <App />
+                </Router>
+            </AppContextProvider>
+        </HelmetProvider>
+    </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
