@@ -1,13 +1,13 @@
-import { ReportHandler } from 'web-vitals';
+import type { Metric } from 'web-vitals';
 
-const reportWebVitals = (onPerfEntry?: ReportHandler) => {
+const reportWebVitals = (onPerfEntry?: (metric: Metric) => void) => {
     if (onPerfEntry && onPerfEntry instanceof Function) {
-        import('web-vitals').then(({ getCLS, getFID, getFCP, getLCP, getTTFB }) => {
-            getCLS(onPerfEntry);
-            getFID(onPerfEntry);
-            getFCP(onPerfEntry);
-            getLCP(onPerfEntry);
-            getTTFB(onPerfEntry);
+        import('web-vitals').then(({ onCLS, onFID, onFCP, onLCP, onTTFB }) => {
+            onCLS(onPerfEntry);
+            onFID(onPerfEntry);
+            onFCP(onPerfEntry);
+            onLCP(onPerfEntry);
+            onTTFB(onPerfEntry);
         });
     }
 };
